requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SocketIO==5.4.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kafka-python==2.0.2
MarkupSafe==3.0.2
netaddr==1.3.0
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
scapy==2.6.0
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
Werkzeug==3.1.0
wsproto==1.2.0
